Rename Barr-Houseman analytic module for 1996 reference
Rename analytic/barr_houseman.py to barr_houseman_96.py so the module path identifies the paper that supplies the closed-form Cartesian solution and cannot be confused with the related BH92 study. Update the package import and developer documentation, and add a focused assertion that uw.analytic.barr_houseman_96.FaultedMedium resolves to the registered public class.
