+++ src/underworld3/analytic/barr_houseman_96.py
